% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_add_tag.R
\name{ss_add_tag}
\alias{ss_add_tag}
\title{Add tag to an order in ShipStation}
\usage{
ss_add_tag(order_number, tag_id = NULL)
}
\arguments{
\item{order_number}{order number (character/scalar)}

\item{tag_id}{tag id (numeric/scalar)}
}
\description{
Add tag to an order in ShipStation
}
