% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_api.R
\name{ss_api}
\alias{ss_api}
\title{Make a ShipStation API call}
\usage{
ss_api(path, fun = httr::GET, ...)
}
\arguments{
\item{path}{path to api endpoint}

\item{fun}{API function (GET or POST)}
}
\description{
Make a ShipStation API call
}
