% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_get_rates.R
\name{ss_get_carrier_rates}
\alias{ss_get_carrier_rates}
\title{Get shipping rates for specific carriers}
\usage{
ss_get_carrier_rates(carrier, ...)
}
\arguments{
\item{...}{payload passed on to API}
}
\description{
Get shipping rates for specific carriers
}
