% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_get_orders.R
\name{ss_get_orders}
\alias{ss_get_orders}
\title{Call the ShipStation Orders API}
\usage{
ss_get_orders(
  order_number = NULL,
  start_date = Sys.Date() - 1,
  end_date = Sys.Date()
)
}
\arguments{
\item{order_number}{Order Number (integer/vector)}

\item{start_date}{first day of data (Date/scalar)}

\item{end_date}{last day of data (Date/scalar)}
}
\description{
Call the ShipStation Orders API
}
