% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_get_rates.R
\name{ss_get_rates}
\alias{ss_get_rates}
\title{Get shipping rates for ALL carriers for a given order}
\usage{
ss_get_rates(order_number = "3023035", carriers = c("stamps_com", "ups"))
}
\arguments{
\item{order_number}{order number (numeric/scalar)}

\item{carriers}{carriers (character/vector)}
}
\description{
Get shipping rates for ALL carriers for a given order
}
