% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_get_shipment_details.R
\name{ss_get_shipment_details}
\alias{ss_get_shipment_details}
\title{Get recent shipment info/details}
\usage{
ss_get_shipment_details(start_date = Sys.Date() - 7,
  end_date = Sys.Date() - 1)
}
\description{
Get recent shipment info/details
}
