% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_get_shipments.R
\name{ss_get_shipments}
\alias{ss_get_shipments}
\title{Call the ShipStation API}
\usage{
ss_get_shipments(
  order_number = NULL,
  start_date = Sys.Date() - 1,
  end_date = Sys.Date() - 1
)
}
\arguments{
\item{order_number}{Order Number (integer/vector)}

\item{start_date}{first day of data (Date/scalar)}

\item{end_date}{last day of data (Date/scalar)}
}
\description{
Call the ShipStation API
}
