% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_get_single_item_shipping_weights.R
\name{ss_get_single_item_shipping_weights}
\alias{ss_get_single_item_shipping_weights}
\title{Estimate product weights based on shipping weights for recent single-item orders}
\usage{
ss_get_single_item_shipping_weights(
  start_date = Sys.Date() - 7,
  end_date = Sys.Date() - 1
)
}
\arguments{
\item{start_date}{first day of data (Date/scalar)}

\item{end_date}{last day of data (Date/scalar)}
}
\description{
Estimate product weights based on shipping weights for recent single-item orders
}
