% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_mark_as_shipped.R
\name{ss_mark_as_shipped}
\alias{ss_mark_as_shipped}
\title{Mark orders as shipped and add tracking info}
\usage{
ss_mark_as_shipped(
  orderid,
  carrier,
  tracking,
  ship_date,
  notify_customer = FALSE,
  notify_channel = TRUE
)
}
\arguments{
\item{orderid}{OrderId}

\item{tracking}{tracking number}

\item{ship_date}{date shipped}

\item{notify_customer}{should customer be notified? (logical/scalar)}

\item{notify_channel}{should channel be notified? (logical/scalar)}

\item{carrient}{shipping carrier}
}
\description{
Mark orders as shipped and add tracking info
}
